# spec/openapi.yaml
openapi: 3.1.0
info:
  title: DeFi Dashboard API
  version: 1.0.0
  description: Backend API for DeFi portfolio tracking and management
  contact:
    name: API Support
    email: api@defi-dashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.defi-dashboard.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - code
        - message
...
    MagicLinkRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    MagicLinkResponse:
      type: object
      properties:
        message:
          type: string
        magicLink:
          type: string
          description: Magic link URL (development only)
      required:
        - message
    Token:
      type: object
      properties:
        address:
          type: string
        symbol:
          type: string
        name:
          type: string
        decimals:
          type: integer
        chainId:
          type: integer
        logoUri:
          type: string
        price:
          type: number
        priceChange24h:
          type: number
      required:
        - address
        - symbol
        - decimals
        - chainId
...
    BridgeRoute:
      type: object
      properties:
        id:
          type: string
        fromChain:
          type: integer
        toChain:
          type: integer
        fromToken:
          $ref: '#/components/schemas/Token'
        toToken:
          $ref: '#/components/schemas/Token'
        bridge:
          type: string
        estimatedTime:
          type: integer
        fee:
          type: number
        feeUsd:
          type: number
      required:
        - id
        - fromChain
        - toChain
        - fromToken
        - toToken
        - bridge
        - estimatedTime
        - fee
...
security:
  - bearerAuth: []
paths:
...
  /auth/magic-link:
    post:
      tags:
        - auth
      summary: Send magic link authentication email (stub)
      description: |
        Stub implementation for email-based magic link authentication.
        In development, returns the magic link URL for testing.
        In production, only sends the email and returns a success message.
      operationId: sendMagicLink
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
            example:
              email: "user@example.com"
      responses:
        '200':
          description: Magic link sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkResponse'
              example:
                message: "Magic link sent to your email"
                magicLink: "https://your-domain.com/auth/verify-magic?token=abc123def456"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /bridge/routes:
    post:
      tags:
        - bridge
      summary: Get bridge routes
      operationId: getBridgeRoutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromChain:
                  type: integer
                toChain:
                  type: integer
                fromToken:
                  type: string
                toToken:
                  type: string
                amount:
                  type: string
              required:
                - fromChain
                - toChain
                - fromToken
                - amount
      responses:
        '200':
          description: Available bridge routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  routes:
                    type: array
                    items:
                      $ref: '#/components/schemas/BridgeRoute'
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "VALIDATION_ERROR"
            message: "Invalid request parameters"
            details:
              field: "address"
              issue: "Invalid Ethereum address format"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred"